// app/(dashboard)/certificates/eicr/[id]/update/schedule-of-inspections/distribution-Circuits/schema.ts
import { z } from "zod";

export const UpdateDistributionCircuitsSchema = z.object({
  id: z.string().cuid(),
  item_5_1: z.string().optional(),
  item_5_2: z.string().optional(),
  item_5_3: z.string().optional(),
  item_5_4: z.string().optional(),
  item_5_5: z.string().optional(),
  item_5_6: z.string().optional(),
  item_5_7: z.string().optional(),
  item_5_8: z.string().optional(),
  item_5_9: z.string().optional(),
  item_5_10: z.string().optional(),
  item_5_11: z.string().optional(),
  item_5_12: z.string().optional(),
  item_5_13: z.string().optional(),
  item_5_14A: z.string().optional(),
  item_5_14B: z.string().optional(),
  item_5_15: z.string().optional(),
  item_5_16: z.string().optional(),
  item_5_17: z.string().optional(),
  item_5_18: z.string().optional(),
  item_5_19: z.string().optional(),
  item_5_20: z.string().optional(),
  item_5_21: z.string().optional(),
  item_5_22: z.string().optional(),
  item_5_23: z.string().optional(),
  item_5_24: z.string().optional(),
});